EDGEPOD-406 : Move nas encoding in the mme-app
Fixing code re-org
