+++ Dockerfile
@@ -1,4 +1,4 @@
-#
+
 # Copyright 2019-present Open Networking Foundation
 #
 # SPDX-License-Identifier: Apache-2.0
@@ -11,8 +11,39 @@ WORKDIR /openmme
 COPY install_builddeps.sh .
 RUN ./install_builddeps.sh
 
-COPY . ./
-RUN make && make install
+#COPY . ./
+COPY Makefile Makefile.common ./
+COPY include/cmn ./include/cmn
+COPY include/common ./include/common
+COPY include/stateMachineFwk ./include/stateMachineFwk
+COPY src/cmn ./src/cmn
+COPY src/common ./src/common
+COPY src/gtpV2Codec ./src/gtpV2Codec
+COPY src/mmeGrpcClient ./src/mmeGrpcClient
+COPY src/mmeGrpcProtos ./src/mmeGrpcProtos
+COPY src/stateMachineFwk ./src/stateMachineFwk
+RUN make -C src/cmn
+RUN make -C src/common
+RUN make -C src/gtpV2Codec
+RUN make -C src/stateMachineFwk
+RUN make -C src/mmeGrpcProtos
+RUN make -C src/mmeGrpcClient
+RUN make -C src/cmn
+COPY include/s11 ./include/s11
+COPY include/s6a ./include/s6a
+COPY include/s1ap ./include/s1ap
+COPY include/mme-app ./include/mme-app
+COPY src/s1ap/asn1c ./src/s1ap/asn1c
+COPY src/s11 ./src/s11
+RUN make -C src/s11
+COPY src/s6a ./src/s6a
+RUN make -C src/s6a
+COPY src/mme-app ./src/mme-app
+RUN make -C src/mme-app
+COPY src/s1ap ./src/s1ap
+RUN make -C src/s1ap
+
+RUN make install
 
 FROM $BASE_OS AS runtime
 COPY install_rundeps.sh .
