+++ Makefile.common
@@ -45,7 +45,16 @@ DEBUG = false
 
 GRPC_ROOT=/tmp/grpc
 
+# to enable NAS message encoding on S1AP enable below flag. In this case
+# mme-app will have to send all the details and then s1ap container will encode
+# NAS message 
 #CFLAGS += -DS1AP_ENCODE_NAS
+
+# to enable NAS message decoding at the S1AP enable following commented line, 
+#this will cause s1ap container to decoee NAS message and forward the decoded
+#messasge content to mme-app
+#CFLAGS += -DS1AP_DECODE_NAS
+
 CFLAGS += -D_FORTIFY_SOURCE=2 -fasynchronous-unwind-tables -fexceptions
 CFLAGS += -fstack-protector-all -fstack-protector-strong -Wall
 CFLAGS += -Werror=format-security -Werror=implicit-function-declaration
