+++ include/common/nas_structs.h
@@ -11,20 +11,24 @@
 #define __NAS_STRUCTS_H_
 #include <stdint.h>
 #include <stdbool.h>
+#include "ProcedureCode.h"
+#include "nas_headers.h"
+
+#define AUTH_REQ_NO_OF_NAS_IES 2
+#define SEC_MODE_NO_OF_NAS_IES 1
+#define ICS_REQ_NO_OF_NAS_IES 5
+#define TAU_RSP_NO_OF_NAS_IES 2
+
+/* ESM messages */
+#define ESM_MSG_ACTV_DEF_BEAR__CTX_REQ 0xc1
+
+#define AUTH_SYNC_FAILURE 21
 
 #include "sec.h"
-#include "ProcedureCode.h"
 #include "Criticality.h"
 #include "S1AP-PDU.h"
-#include "InitiatingMessage.h"
 
 #define MAX_NAS_MSG_SIZE 500
-#define BUFFER_SIZE 255
-
-typedef struct Buffer {
-	unsigned char buf[BUFFER_SIZE];
-	unsigned char pos;
-}Buffer;
 
 /*TODO: Needed fro paging. Ignoring for now*/
 struct proto_conf {
@@ -40,53 +44,6 @@ struct pco
     unsigned char pco_length;
     unsigned char pco_options[MAX_PCO_OPTION_SIZE];
 };
-
-/*NAS MSG IE CODES */
-/* Message content : 
-   3gpp 24.301
-   Table 8.2.4.1: IEI Column.*/
-typedef enum
-{
-    NAS_IE_TYPE_EPS_MOBILE_ID_IMSI=0x1,
-    NAS_IE_TYPE_UE_NETWORK_CAPABILITY=0x2,
-    NAS_IE_TYPE_ESM_MSG=0x3,
-    NAS_IE_TYPE_TMSI_STATUS=0x09,
-    NAS_IE_TYPE_MS_NETWORK_FEATURE_SUPPORT=0x0C,
-    NAS_IE_TYPE_GUTI_TYPE=0x0E,
-    NAS_IE_TYPE_ADDITIONAL_UPDATE_TYPE=0xF,
-    NAS_IE_TYPE_MS_CLASSMARK_2=0x11,
-    NAS_IE_TYPE_LAI=0x13,
-    NAS_IE_TYPE_PTMSI_SIGNATURE=0x19,
-    NAS_IE_TYPE_MS_CLASSMARK_3=0x20,
-    NAS_IE_TYPE_APN=0x28,
-    NAS_IE_TYPE_AUTH_FAIL_PARAM=0x30,
-    NAS_IE_TYPE_MS_NETWORK_CAPABILITY=0x31,
-    NAS_IE_TYPE_DRX_PARAM=0x5C,
-    NAS_IE_TYPE_TAI=0x52,
-    NAS_IE_TYPE_VOICE_DOMAIN_PREF_UE_USAGE_SETTING=0x5D,
-    NAS_IE_TYPE_TX_FLAG=0xAA,
-    NAS_IE_TYPE_PCO=0xAB,
-    NAS_IE_TYPE_PTI=0xAC,
-}nas_ie_type;
-
-typedef struct allocation_retention_prio {
-	uint8_t prioLevel :4;
-	uint8_t preEmptionCapab :1;
-	uint8_t preEmptionVulnebility :1;
-	uint8_t spare :2;
-} allocation_retention_prio;
-
-
-typedef struct E_RAB_Level_QoS_Params {
-	uint8_t qci;
-	//uint8_t ext;
-    allocation_retention_prio arPrio;
-} E_RAB_Level_QoS_Params;
-
-typedef struct ue_aggregate_maximum_bitrate {
-	uint32_t uEaggregateMaxBitRateDL;
-	uint32_t uEaggregateMaxBitRateUL;
-} ue_aggregate_maximum_bitrate;
 
 typedef struct pdn_address {
 	uint8_t spare :5;
@@ -104,20 +61,6 @@ enum protocol_discriminator{
     EPSMobilityManagementMessages = 7,
 };
 	
-enum eps_nas_mesage_type {
-	AttachAccept = 0x42,
-	AttachReject = 0x44,
-	DetachAccept = 0x46,
-	DetachRequest = 0x45,
-	TauAccept    = 0x49,
-    TauReject    = 0x4b,
-	ServiceReject = 0x4e,
-	AuthenticationRequest = 0x52,
-    IdentityRequest       = 0x55,
-	SecurityModeCommand = 0x5d,
-    EMMInformation = 0x61,
-	ESMInformationRequest = 0xd9,
-};
 
 typedef struct esm_qos
 {
@@ -163,69 +106,17 @@ typedef struct ESM_APN_AMBR {
 	/* uint8_t ul_total; */
 } ESM_APN_AMBR;
 
-enum procedure_code {
-	id_InitialContextSetup = 9,
-	id_downlinkNASTransport = 11,
-	id_errorIndication = 15,
-	id_UEContexRelease = 23,
-};
-
 #define NAS_MSG_UE_IE_GUTI  0x00000001
 #define NAS_MSG_UE_IE_IMSI  0x00000002
 #define UE_ID_IMSI(flags)   ((flags & NAS_MSG_UE_IE_IMSI) == NAS_MSG_UE_IE_IMSI)
 #define UE_ID_GUTI(flags)   ((flags & NAS_MSG_UE_IE_GUTI) == NAS_MSG_UE_IE_GUTI)
 
 
-#define AUTH_REQ_NO_OF_IES 3
-#define SEC_MODE_NO_OF_IES 3
-#define ESM_REQ_NO_OF_IES 3
-#define EMM_INFO_REQ_NO_OF_IES 3
-#define ICS_REQ_NO_OF_IES 6
-#define DTCH_ACCEPT_NO_OF_IES 3
-#define UE_CTX_RELEASE_NO_OF_IES 3
-#define ATTACH_REJECT_NO_OF_IES 3
-#define SERVICE_REJECT_NO_OF_IES 3 
-#define ATTACH_ID_REQUEST_NO_OF_IES 3
-#define TAU_RSP_NO_OF_IES 3
-#define NI_UE_CTX_RELEASE_NO_OF_IES 3
-#define NI_DTCH_REQUEST_NO_OF_IES 3
-#define AUTH_REQ_NO_OF_NAS_IES 2
-#define SEC_MODE_NO_OF_NAS_IES 1
-#define ICS_REQ_NO_OF_NAS_IES 5
-#define TAU_RSP_NO_OF_NAS_IES 2
 
 #define AUTHREQ_NAS_SECURITY_PARAM 0x01
 #define SERVICE_REQ_SECURITY_HEADER 12
-#define AUTH_SYNC_FAILURE 21
-#define AUTH_RESPONSE 53
-
-enum criticality{
-	CRITICALITY_REJECT = 0x0,
-	CRITICALITY_IGNORE = 0x40,
-	CRITICALITY_NOTIFY,
-};
-
-enum protocolie_id {
-	id_MME_UE_S1AP_ID = 0,
-	id_Cause = 2,
-	id_eNB_UE_S1AP_ID = 8,
-	id_ERABToBeSetupListCtxtSUReq = 24,
-	id_NAS_PDU = 26,
-	id_ERABToBeSetupItemCtxtSUReq = 52,
-	id_uEaggregatedMaximumBitrate = 66,
-	id_SecurityKey = 73,
-	id_UE_S1AP_IDs = 99,
-	id_UESecurityCapabilities = 107,
-};
-
-
-/* ESM messages */
-#define ESM_MSG_ACTV_DEF_BEAR__CTX_REQ 0xc1
-#define MSISDN_STR_LEN 10
 
 #define BINARY_IMSI_LEN 8 /*same as packet capture. TODO: Write macros*/
-#define MSISDN_STR_LEN 10
-#define BCD_IMSI_STR_LEN 15
 
 /* NAS Security Encryption Algorithm */
 typedef enum security_encryption_algo {
@@ -258,65 +149,7 @@ typedef enum emmCause {
     emmCause_network_failure = 11
 } e_emmCause;
 
-/* Dependencies */
-typedef enum s1apCause_PR {
-    s1apCause_PR_NOTHING,   /* No components present */
-    s1apCause_PR_radioNetwork,
-    s1apCause_PR_transport,
-    s1apCause_PR_nas,
-    s1apCause_PR_protocol,
-    s1apCause_PR_misc
-
-} s1apCause_PR;
-
-typedef enum s1apCause_PR_transporauseRadioNetwork {
-    s1apCauseRadioNetwork_unspecified   = 0,
-    s1apCauseRadioNetwork_tx2relocoverall_expiry    = 1,
-    s1apCauseRadioNetwork_successful_handover   = 2,
-    s1apCauseRadioNetwork_release_due_to_eutran_generated_reason    = 3,
-    s1apCauseRadioNetwork_handover_cancelled    = 4,
-    s1apCauseRadioNetwork_partial_handover  = 5,
-    s1apCauseRadioNetwork_ho_failure_in_target_EPC_eNB_or_target_system = 6,
-    s1apCauseRadioNetwork_ho_target_not_allowed = 7,
-    s1apCauseRadioNetwork_tS1relocoverall_expiry    = 8,
-    s1apCauseRadioNetwork_tS1relocprep_expiry   = 9,
-    s1apCauseRadioNetwork_cell_not_available    = 10,
-    s1apCauseRadioNetwork_unknown_targetID  = 11,
-    s1apCauseRadioNetwork_no_radio_resources_available_in_target_cell   = 12,
-    s1apCauseRadioNetwork_unknown_mme_ue_s1ap_id    = 13,
-    s1apCauseRadioNetwork_unknown_enb_ue_s1ap_id    = 14,
-    s1apCauseRadioNetwork_unknown_pair_ue_s1ap_id   = 15,
-    s1apCauseRadioNetwork_handover_desirable_for_radio_reason   = 16,
-    s1apCauseRadioNetwork_time_critical_handover    = 17,
-    s1apCauseRadioNetwork_resource_optimisation_handover    = 18,
-    s1apCauseRadioNetwork_reduce_load_in_serving_cell   = 19,
-    s1apCauseRadioNetwork_user_inactivity   = 20,
-    s1apCauseRadioNetwork_radio_connection_with_ue_lost = 21,
-    s1apCauseRadioNetwork_load_balancing_tau_required   = 22,
-    s1apCauseRadioNetwork_cs_fallback_triggered = 23,
-    s1apCauseRadioNetwork_ue_not_available_for_ps_service   = 24,
-    s1apCauseRadioNetwork_radio_resources_not_available = 25,
-    s1apCauseRadioNetwork_failure_in_radio_interface_procedure  = 26,
-    s1apCauseRadioNetwork_invalid_qos_combination   = 27,
-    s1apCauseRadioNetwork_interrat_redirection  = 28,
-    s1apCauseRadioNetwork_interaction_with_other_procedure  = 29,
-    s1apCauseRadioNetwork_unknown_E_RAB_ID  = 30,
-    s1apCauseRadioNetwork_multiple_E_RAB_ID_instances   = 31,
-    s1apCauseRadioNetwork_encryption_and_or_integrity_protection_algorithms_not_supported   = 32,
-    s1apCauseRadioNetwork_s1_intra_system_handover_triggered    = 33,
-    s1apCauseRadioNetwork_s1_inter_system_handover_triggered    = 34,
-    s1apCauseRadioNetwork_x2_handover_triggered = 35,
-    s1apCauseRadioNetwork_redirection_towards_1xRTT = 36,
-    s1apCauseRadioNetwork_not_supported_QCI_value   = 37,
-    s1apCauseRadioNetwork_invalid_CSG_Id    = 38,
-    s1apCauseRadioNetwork_release_due_to_pre_emption    = 39
-} e_s1apCauseRadioNetwork;
-
-typedef enum s1apCauseTransport {
-    s1apCauseTransport_transport_resource_unavailable   = 0,
-    s1apCauseTransport_unspecified  = 1
-} e_s1apCauseTransport;
-
+//this should be nas specifi ?
 typedef enum s1apCauseNas {
     s1apCauseNas_normal_release = 0,
     s1apCauseNas_authentication_failure = 1,
@@ -325,62 +158,10 @@ typedef enum s1apCauseNas {
     s1apCauseNas_csg_subscription_expiry    = 4
 } e_s1apCauseNas;
 
-typedef enum s1apCauseProtocol {
-    s1apCauseProtocol_transfer_syntax_error = 0,
-    s1apCauseProtocol_abstract_syntax_error_reject  = 1,
-    s1apCauseProtocol_abstract_syntax_error_ignore_and_notify   = 2,
-    s1apCauseProtocol_message_not_compatible_with_receiver_state    = 3,
-    s1apCauseProtocol_semantic_error    = 4,
-    s1apCauseProtocol_abstract_syntax_error_falsely_constructed_message = 5,
-    s1apCauseProtocol_unspecified   = 6
-} e_s1apCauseProtocol;
-
-typedef enum s1apCauseMisc {
-    s1apCauseMisc_control_processing_overload   = 0,
-    s1apCauseMisc_not_enough_user_plane_processing_resources    = 1,
-    s1apCauseMisc_hardware_failure  = 2,
-    s1apCauseMisc_om_intervention   = 3,
-    s1apCauseMisc_unspecified   = 4,
-    s1apCauseMisc_unknown_PLMN  = 5
-} e_s1apCauseMisc;
-
-
-typedef struct s1apCause {
-    s1apCause_PR present;
-    union s1apCause_u {
-        s1apCauseRadioNetwork_t  radioNetwork;
-        s1apCauseTransport_t     transport;
-        s1apCauseNas_t   nas;
-        s1apCauseProtocol_t  protocol;
-        s1apCauseMisc_t  misc;
-    } choice;
-} s1apCause_t;
-
-/*36.413: 9.2.1.3a*/
-enum ie_RRC_est_cause {
-	EMERGENCY,
-	HIGHPRIORITYACCESS,
-	MT_ACCESS,
-	MO_SIGNALLING,
-	MO_DATA,
-	DELAYTOLERANTACCESS,
-	MO_VOICECALL,
-	MO_EXCEPTIONDATA
-};
-
-
 /*36.413: 9.2.3.8 - MCC, MCN : Only 3 bytes are used*/
 struct PLMN {
 	unsigned char  idx[3];
 };
-
-//Service Request
-struct STMSI {
-	uint8_t mme_code;
-        uint32_t m_TMSI;
-
-};
-
 
 typedef struct guti {
 	uint8_t spare :4;
@@ -421,23 +202,9 @@ APN name can be in range of min 3 octets to max 102 octets
 */
 #define MAX_APN_LEN 102
 
-
 struct apn_name {
 	unsigned char len;
 	unsigned char val[MAX_APN_LEN];/*TODO: Make dynamic as range is big 3-102*/
-};
-/*36.413: 9.1.8.4*/
-#define ENB_NAME_SIZE 150
-struct ie_enb_name {
-	char enb_name[ENB_NAME_SIZE];
-};
-
-/*36.413: 9.2.1.37*/
-#define MACRO_ENB_ID_SIZE 20
-struct ie_global_enb_id {
-	int plmn;
-	char macro_enb_id[MACRO_ENB_ID_SIZE];
-	/*TODO: make union of enb IDs*/
 };
 
 #define MAC_SIZE 4
@@ -532,7 +299,10 @@ typedef struct nas_pdu_elements {
    nas_pdu_elements_union pduElement;
 }nas_pdu_elements;
 
-
+#define NAS_MSG_UE_IE_GUTI  0x00000001
+#define NAS_MSG_UE_IE_IMSI  0x00000002
+#define UE_ID_IMSI(flags)   ((flags & NAS_MSG_UE_IE_IMSI) == NAS_MSG_UE_IE_IMSI)
+#define UE_ID_GUTI(flags)   ((flags & NAS_MSG_UE_IE_GUTI) == NAS_MSG_UE_IE_GUTI)
 typedef struct nasPDU {
 	nas_pdu_header header;
 	unsigned char elements_len;
@@ -541,149 +311,38 @@ typedef struct nasPDU {
 } nasPDU;
 
 
-typedef struct ue_sec_capabilities {
-	uint8_t eea1 :1;
-	uint8_t eea2 :1;
-	uint8_t eea3 :1;
-	uint8_t eea4 :1;
-	uint8_t eea5 :1;
-	uint8_t eea6_128 :1;
-	uint8_t eea7_128 :1;
-	uint8_t eea8 :1;
-
-	uint8_t eia1 :1;
-	uint8_t eia2 :1;
-	uint8_t eia3 :1;
-	uint8_t eia4 :1;
-	uint8_t eia5 :1;
-	uint8_t eia6_128 :1;
-	uint8_t eia8_128 :1;
-	uint8_t eia8 :1;
-
-	uint8_t uea1 :1;
-	uint8_t uea2 :1;
-	uint8_t uea3 :1;
-	uint8_t uea4 :1;
-	uint8_t uea5 :1;
-	uint8_t uea6 :1;
-	uint8_t uea7 :1;
-	uint8_t uea8 :1;
-
-	uint8_t uia1 :1;
-	uint8_t uia2 :1;
-	uint8_t uia3 :1;
-	uint8_t uia4 :1;
-	uint8_t uia5 :1;
-	uint8_t uia6 :1;
-	uint8_t uia7 :1;
-	uint8_t spare1 :1;
-
-	uint8_t gea1 :1;
-	uint8_t gea2 :1;
-	uint8_t gea3 :1;
-	uint8_t gea4 :1;
-	uint8_t gea5 :1;
-	uint8_t gea6 :1;
-	uint8_t gea7 :1;
-	uint8_t spare2 :1;
-} ue_sec_capabilities;
-
-typedef struct ERABSetup {
-	uint8_t e_RAB_ID;
-	E_RAB_Level_QoS_Params e_RAB_QoS_Params;
-	uint32_t transportLayerAddress;
-	uint32_t gtp_teid;
-	struct nasPDU nas;
-} ERABSetup;
-
-/**E-RAB structure declarations**/
-struct eRAB_header { //TODO: making provision, chec -is it needed?
-	unsigned char criticality;
-};
-
-struct eRAB_setup_ctx_SU {
-	unsigned short eRAB_id;
-	unsigned short dont_know_byte;
-	unsigned int transp_layer_addr;
-	unsigned int gtp_teid;
-};
-
-union eRAB_IE {
-	struct eRAB_setup_ctx_SU su_res;
-};
-
-typedef struct eRAB_elements {
-	struct eRAB_header header;
-	unsigned char no_of_elements;
-	union eRAB_IE *elements;
-}eRAB_elements;
-
-
-#define SECURITY_KEY_SIZE 32
-
-typedef struct proto_IE_data {
-	int 			IE_type;
-    union value{
-        struct ie_enb_name 	enb_name;
-        struct ie_global_enb_id global_enb_id;
-        long			enb_ue_s1ap_id;
-        long			mme_ue_s1ap_id;
-        struct 			nasPDU nas;
-        struct s1apCause cause;
-        struct TAI 		 tai;
-        struct CGI 		 utran_cgi;
-        struct STMSI	 s_tmsi;
-        enum ie_RRC_est_cause 	rrc_est_cause;
-        struct eRAB_elements 	erab;
-        ue_aggregate_maximum_bitrate ue_aggrt_max_bit_rate;
-        ERABSetup E_RABToBeSetupItemCtxtSUReq;
-        ue_sec_capabilities ue_sec_capab;
-        uint8_t sec_key[SECURITY_KEY_SIZE];
-    }val;
-}proto_IEs;
-
-struct proto_IE {
-    ProcedureCode_t  procedureCode;
-    Criticality_t    criticality;
-	short 		no_of_IEs;
-	proto_IEs	*data;
-	uint8_t     ie_nas_index;
-    uint8_t     ie_tai_index;
-    uint8_t     ie_cgi_index;
-};
-
-	/* NAS Security Header */
-	typedef enum security_header_type {
-	    Plain = 0,
-	    IntegrityProtected,
-	    IntegrityProtectedCiphered,
-	    IntegrityProtectedEPSSecCntxt,
-	}security_header_type;
+/* NAS Security Header */
+typedef enum security_header_type {
+    Plain = 0,
+    IntegrityProtected,
+    IntegrityProtectedCiphered,
+    IntegrityProtectedEPSSecCntxt,
+}security_header_type;
 	
-	typedef struct nas_pdu_header_sec {
-	        unsigned char security_header_type:4;
-	        unsigned char proto_discriminator:4;
-	        unsigned char mac[MAC_SIZE];
-	        unsigned char seq_no;
-	}nas_pdu_header_sec;
+typedef struct nas_pdu_header_sec {
+        unsigned char security_header_type:4;
+        unsigned char proto_discriminator:4;
+        unsigned char mac[MAC_SIZE];
+        unsigned char seq_no;
+}nas_pdu_header_sec;
 	
-	typedef struct nas_pdu_header_short {
-	        unsigned char security_header_type:4;
-	        unsigned char proto_discriminator:4;
-	        unsigned char message_type;
-	}nas_pdu_header_short;
-	
-	typedef struct nas_pdu_header_long {
-	        unsigned char security_header_type:4;
-	        unsigned char proto_discriminator:4;
-	        unsigned char procedure_trans_identity;
-	        unsigned char message_type;
-	}nas_pdu_header_long;
+typedef struct nas_pdu_header_short {
+        unsigned char security_header_type:4;
+        unsigned char proto_discriminator:4;
+        unsigned char message_type;
+}nas_pdu_header_short;
 
-typedef struct s1ap_PDU {
-	unsigned char procedurecode;
-	unsigned char criticality;
-	struct proto_IE value;
-}s1ap_PDU;
+typedef struct nas_pdu_header_long {
+        unsigned char security_header_type:4;
+        unsigned char proto_discriminator:4;
+        unsigned char procedure_trans_identity;
+        unsigned char message_type;
+}nas_pdu_header_long;
+
+enum ie_fail_internal_cause {
+	AIA_FAIL,
+	SECURITY_MOD_REJECT_FAIL,
+	MAX_FAIL_CAUSE
+};
 
 #endif /*__S1AP_STRUCTS_H*/
