+++ include/s1ap/s1ap_msg_codes.h
@@ -37,23 +37,4 @@
 #define S1AP_IE_RRC_EST_CAUSE  134
 #define S1AP_ERAB_SETUP_CTX_SUR 51
 
-/* should be part of mme app only */
-/*NAS message type codes*/
-#define NAS_ESM_RESP 0xda
-#define NAS_AUTH_RESP 0x53
-#define NAS_AUTH_REJECT 0x54
-#define NAS_AUTH_FAILURE 0x5c
-#define NAS_IDENTITY_REQUEST 0x55
-#define NAS_IDENTITY_RESPONSE 0x56
-#define NAS_SEC_MODE_COMPLETE 0x5e
-#define NAS_SEC_MODE_REJECT  0x5f
-#define NAS_ATTACH_REQUEST 0x41
-#define NAS_ATTACH_COMPLETE 0x43
-#define NAS_ATTACH_REJECT 0x44
-#define NAS_TAU_REQUEST    0x48
-#define NAS_TAU_COMPLETE   0x4a
-#define NAS_DETACH_REQUEST 0x45
-#define NAS_DETACH_ACCEPT 0x46
-#define NAS_SERVICE_REQUEST 0x4D
-
 #endif /*__S1AP_MSG_CODES*/
