+++ src/s1ap/handlers/attach_authresp.c
@@ -53,6 +53,7 @@ s1_auth_resp_handler(struct proto_IE *s1_auth_resp_ies)
                 {
 	                auth_resp.ue_idx = s1_auth_resp_ies->data[i].val.mme_ue_s1ap_id;
                 }break;
+#ifdef S1AP_DECODE_NAS
             case S1AP_IE_NAS_PDU:
                 {
                     if(s1_auth_resp_ies->data[i].val.nas.header.message_type != NAS_AUTH_RESP)
@@ -69,6 +70,7 @@ s1_auth_resp_handler(struct proto_IE *s1_auth_resp_ies)
                            sizeof(struct XRES));
 
                 }break;
+#endif
             default:
                 log_msg(LOG_WARNING,"Unhandled IE %d \n",s1_auth_resp_ies->data[i].IE_type);
         }
@@ -116,6 +118,7 @@ s1_auth_fail_handler(struct proto_IE *s1_auth_resp_ies)
                 {
 	                auth_resp.ue_idx = s1_auth_resp_ies->data[i].val.mme_ue_s1ap_id;
                 }break;
+#ifdef S1AP_DECODE_NAS
             case S1AP_IE_NAS_PDU:
                 {
                     auth_resp.msg_data.authresp_Q_msg_m.status = S1AP_AUTH_FAILED;//Error in authentication
@@ -123,6 +126,7 @@ s1_auth_fail_handler(struct proto_IE *s1_auth_resp_ies)
                            &(s1_auth_resp_ies->data[i].val.nas.elements[0].pduElement.auth_fail_resp),
 		                   sizeof(struct AUTS));
                 }break;
+#endif
             default:
                 log_msg(LOG_WARNING,"Unhandled IE %d \n",s1_auth_resp_ies->data[i].IE_type);
         }
