+++ src/s1ap/handlers/attach_esmresp.c
@@ -49,6 +49,7 @@ s1_esm_resp_handler(struct proto_IE *s1_esm_resp_ies)
                 {
 	                esm_resp.ue_idx = s1_esm_resp_ies->data[i].val.mme_ue_s1ap_id;
                 }break;
+#ifdef S1AP_DECODE_NAS
             case S1AP_IE_NAS_PDU:
                 {
                     if(s1_esm_resp_ies->data[i].val.nas.header.message_type != NAS_ESM_RESP)
@@ -62,6 +63,7 @@ s1_esm_resp_handler(struct proto_IE *s1_esm_resp_ies)
 		                       sizeof(struct apn_name));
                     }
                 }break;
+#endif
             default:
                 log_msg(LOG_WARNING,"Unhandled IE %d \n",s1_esm_resp_ies->data[i].IE_type);
         }
