+++ src/s1ap/handlers/service_req.c
@@ -64,6 +64,7 @@ s1_init_ue_service_req_handler(struct proto_IE *service_req_ies, int enb_fd)
                     log_msg(LOG_INFO, "Service Req S1AP_IE_ENB_UE_ID.\n");
                     req.s1ap_enb_ue_id = service_req_ies->data[i].val.enb_ue_s1ap_id;
                 }break;
+#ifdef S1AP_DECODE_NAS
             case S1AP_IE_NAS_PDU:
                 {
                     log_msg(LOG_INFO, "Service Req NAS PDU.\n");
@@ -71,6 +72,7 @@ s1_init_ue_service_req_handler(struct proto_IE *service_req_ies, int enb_fd)
                     req.msg_data.service_req_Q_msg_m.seq_no = service_req_ies->data[i].val.nas.header.seq_no;
                     memcpy(&req.msg_data.service_req_Q_msg_m.mac, service_req_ies->data[i].val.nas.header.short_mac, sizeof(uint16_t));
                 }break;
+#endif
             case S1AP_IE_TAI:
                 {
                     log_msg(LOG_INFO, "Service Req TAI.\n");
